moved functionality so that changecolor component works with colorlist
